Avoid making another HTTP request
To comply with the Pyfilesystem exceptions we must verify that the thing we listed was indeed a directory. We can infer this information from the results of our WebHDFS query, so there's no need to query again. This essentially cuts the number of web requests (and time) in half for recursive directory walking.