Defaulted hdfs to using hostname and port from env
If not specified in the URI we will default to using the hostname as
defined in PYFS_HADOOP_NAMENODE_ADDR. If not specified in the URI we
will default the port number to PYFS_HADOOP_NAMENODE_PORT or 50070.

The following URIs can now be opened, where the env defines the host:

    hdfs://
    hdfs:///
    hdfs:///subfolder
    hdfs://:80
    hdfs://:80/
    hdfs://host:
    hdfs://host:/
    hdfs://host:/subfolder
    hdfs://host:80
    hdfs://host:80/
    hdfs://host:80/subfolder

Naice.